chore: rename package configure -> snap
