+++ src/main/java/com/canonical/devpackspring/snap/Manifest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.canonical.devpackspring.configure;
+package com.canonical.devpackspring.snap;
 
 import java.io.ByteArrayInputStream;
 import java.io.File;
