+++ src/test/java/com/canonical/devpackspring/snap/Util.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.canonical.devpackspring.configure;
+package com.canonical.devpackspring.snap;
 
 import java.io.BufferedReader;
 import java.io.IOException;
