+++ src/test/resources/com/canonical/devpackspring/snap/mainfest.yaml
