+++ src/test/resources/com/canonical/devpackspring/snap/settings.xml
